backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.4
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt<5
python-multipart==0.0.20
scikit-learn==1.7.1
numpy==2.3.2
pandas==2.3.2
joblib==1.5.1
shap==0.48.0
transformers==4.55.4
pytest==8.4.1
httpx==0.28.1
ruff==0.12.9
black==25.1.0
isort==6.0.1
